Make eepdump more robust to random files
by not segfaulting when pointing this to random files (like other
EEPROMs in the system, which might have other data in different formats
in them).

This now checks signature and verison, and keeps track of file length
to make sure things are correct.

tested on files created with eepmake, on disk and pointing to EEPROMs.

Signed-off-by: Robin Getz <robin.getz@analog.com>